fix: Implemented check for the existence of review by tourist id and tour id

+++ src/Modules/Tours/Explorer.Tours.API/Dtos/ReviewDto.cs
@@ -11,7 +11,7 @@ namespace Explorer.Tours.API.Dtos
         public int Id { get; set; }
         public int Rating { get; set; }
         public string Comment { get; set; }
-        public int TouristId { get; set; }
+        public long TouristId { get; set; }
         public DateOnly TourVisitDate { get; set; }
         public DateOnly CommentDate { get; set; }
         public int TourId { get; set; }
